+++ infrastructure/repositories/PgStockRepository.ts
@@ -3,7 +3,6 @@ import { Stock } from '@prisma/client';
 import { prisma } from "@/config/prismaClient";
 
 export class PrStockRepository implements IStockRepository {
-  // 주식 코드에 따른 주식 정보 조회
   async findStockByCode(stockCode: string): Promise<Stock | null>  {
     const stock = await prisma.stock.findFirst({
       where: { stockCode: stockCode },
